fix: Pagination styles [MDS-549] (#2252)
