+++ workspaces/core/src/pagination/private/types/ChildrenFunc.ts
@@ -1,0 +1,5 @@
+interface ChildrenFunc {
+  (data: { disabled?: boolean }): React.ReactNode;
+}
+
+export default ChildrenFunc;
