+++ workspaces/core/src/pagination/private/types/NextPrevButtonProps.ts
@@ -1,0 +1,9 @@
+import type ChildrenFunc from './ChildrenFunc';
+
+interface NextPrevButtonProps<C extends React.ElementType> {
+  as?: C;
+  className?: string;
+  children?: React.ReactNode | ChildrenFunc;
+}
+
+export default NextPrevButtonProps;
