+++ workspaces/core/src/pagination/private/types/PagesProps.ts
@@ -1,0 +1,7 @@
+type PagesProps<C extends React.ElementType> = {
+  as?: C;
+  className?: string;
+  truncableText?: JSX.Element | string;
+};
+
+export default PagesProps;
