+++ workspaces/core/src/pagination/private/types/PaginationContext.ts
@@ -1,0 +1,20 @@
+import { createContext } from 'react';
+import type UsePagination from './UsePagination';
+
+const defaultState: UsePagination = {
+  currentPage: 0,
+  setCurrentPage: () => {},
+  pages: [],
+  hasPreviousPage: false,
+  hasNextPage: false,
+  previousPages: [],
+  isPreviousTruncable: false,
+  middlePages: [],
+  isNextTruncable: false,
+  nextPages: [],
+};
+
+const PaginationContext = createContext(defaultState);
+PaginationContext.displayName = 'PaginationContext';
+
+export default PaginationContext;
