+++ workspaces/core/src/pagination/private/types/PaginationProps.ts
@@ -1,0 +1,10 @@
+type PaginationProps = {
+  className?: string;
+  currentPage: number;
+  setCurrentPage: (page: number) => void;
+  totalPages: number;
+  edgePageCount?: number;
+  middlePagesSiblingCount?: number;
+};
+
+export default PaginationProps;
