+++ workspaces/core/src/pagination/private/types/PolymorphicNextPrevButtonProps.ts
@@ -1,0 +1,8 @@
+import type NextPrevButtonProps from './NextPrevButtonProps';
+import type PropsWithChildrenFunc from './PropsWithChildrenFunc';
+
+type PolymorphicNextPrevButtonProps<C extends React.ElementType> =
+  PropsWithChildrenFunc<NextPrevButtonProps<C>> &
+    Omit<React.ComponentPropsWithoutRef<C>, keyof NextPrevButtonProps<C>>;
+
+export default PolymorphicNextPrevButtonProps;
