+++ workspaces/core/src/pagination/private/types/PolymorphicPagesProps.ts
@@ -1,0 +1,7 @@
+import type PagesProps from './PagesProps';
+
+type PolymorphicPagesProps<C extends React.ElementType> =
+  React.PropsWithChildren<PagesProps<C>> &
+    Omit<React.ComponentPropsWithoutRef<C>, keyof PagesProps<C>>;
+
+export default PolymorphicPagesProps;
