+++ workspaces/core/src/pagination/private/types/PropsWithChildrenFunc.ts
@@ -1,0 +1,7 @@
+import type ChildrenFunc from './ChildrenFunc';
+
+type PropsWithChildrenFunc<P = unknown> = P & {
+  children?: React.ReactNode | ChildrenFunc | undefined;
+};
+
+export default PropsWithChildrenFunc;
