+++ workspaces/core/src/pagination/private/types/TruncableElementProps.ts
@@ -1,0 +1,5 @@
+type TruncableElementProps = {
+  prev?: boolean;
+};
+
+export default TruncableElementProps;
