+++ workspaces/core/src/pagination/private/types/UsePagination.ts
@@ -1,0 +1,14 @@
+type UsePagination = {
+  currentPage: number;
+  setCurrentPage: (page: number) => void;
+  pages: number[];
+  hasPreviousPage: boolean;
+  hasNextPage: boolean;
+  previousPages: number[];
+  isPreviousTruncable: boolean;
+  middlePages: number[];
+  isNextTruncable: boolean;
+  nextPages: number[];
+};
+
+export default UsePagination;
