+++ workspaces/core/src/pagination/private/types/WithChildren.ts
@@ -1,0 +1,3 @@
+type WithChildren<T = {}> = T & { children?: React.ReactNode };
+
+export default WithChildren;
